+++ contract/pkg/bucket/cdn_node_param.go
@@ -1,13 +1,15 @@
 package bucket
 
-import "encoding/json"
+import (
+	"encoding/json"
+)
 
 // Structure-helper for json on the CDN Node Params string
 type CDNNodeParams struct {
-	Url       string `json:"url"`
-	Size      uint8  `json:"size"`
-	Location  string `json:"location"`
-	PublicKey string `json:"publicKey"`
+	Url       string  `json:"url"`
+	Size      FlexInt `json:"size"`
+	Location  string  `json:"location"`
+	PublicKey string  `json:"publicKey"`
 }
 
 func ReadCDNNodeParams(s string) (p CDNNodeParams, err error) {