+++ contract/pkg/bucket/cdn_node_param_test.go
@@ -27,6 +27,17 @@ func TestReadCDNNodeParams(t *testing.T) {
 			wantErr: false,
 		},
 		{
+			name: "Positive test with size as string",
+			args: args{s: `{"url":"http://localhost:8080", "publicKey": "NodePublicKey", "size":"1", "location":"US"}`},
+			wantP: CDNNodeParams{
+				Url:       "http://localhost:8080",
+				Size:      1,
+				Location:  "US",
+				PublicKey: "NodePublicKey",
+			},
+			wantErr: false,
+		},
+		{
 			name:    "Wrong structure",
 			args:    args{s: `{"}`},
 			wantP:   CDNNodeParams{},