+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 2. The functionality to send acknowledgements has been implemented.
 3. Subscription to smart contract events has been incorporated to facilitate cache purging.
 4. A new event `BucketAvailabilityUpdated` has been introduced into the contract.
+
+### Bug Fixes
+1. The bug, that ignore replication factor from smart contract if it set as string